fix: fail open on bridge ttyd html read errors
