+++ packages/web/src/app/api/sessions/[id]/terminal/ttyd/route.ts
@@ -10,6 +10,7 @@ import {
   injectTtydResizeShim,
   resolveBridgeSessionTarget,
 } from "@/lib/bridgeTtyd";
+import { readTtydHtmlResponse } from "@/lib/ttydHtmlResponse";
 
 export const dynamic = "force-dynamic";
 export const runtime = "nodejs";
@@ -18,43 +19,6 @@ type RouteContext = {
   params: Promise<{ id: string }>;
 };
 
-const MAX_TTYD_HTML_RESPONSE_BYTES = 512 * 1024;
-
-async function readTtydHtmlResponse(proxied: Response): Promise<string | null> {
-  const contentType = proxied.headers.get("content-type")?.toLowerCase() ?? "";
-  if (!contentType.startsWith("text/html")) {
-    return null;
-  }
-
-  const contentLength = Number.parseInt(proxied.headers.get("content-length") ?? "", 10);
-  if (Number.isFinite(contentLength) && contentLength > MAX_TTYD_HTML_RESPONSE_BYTES) {
-    throw new Error("ttyd frontend response is too large");
-  }
-
-  const reader = proxied.body?.getReader();
-  if (!reader) {
-    return null;
-  }
-
-  const decoder = new TextDecoder();
-  let totalBytes = 0;
-  let html = "";
-
-  for (;;) {
-    const { done, value } = await reader.read();
-    if (done) {
-      break;
-    }
-    totalBytes += value.byteLength;
-    if (totalBytes > MAX_TTYD_HTML_RESPONSE_BYTES) {
-      throw new Error("ttyd frontend response is too large");
-    }
-    html += decoder.decode(value, { stream: true });
-  }
-
-  html += decoder.decode();
-  return html;
-}
 
 /**
  * Inject the resize coordination shim into a proxied HTML response.
